♻️ Refactor: final이랑 final/check이랑 리팩토링
